Remgrouped container/location methods in core.goods
No need for them to be in the separate core.location module now that
it's the same Model

Also regrouped tests in the goods.tests package, but kept the ones
related to containers methods in a separate module.

Finally, referencing these new methods of the Goods model in the apidoc was
the occasion to reference some forgotten Property related methods.
