+++ anyblok_wms_base/core/goods/tests/test_containers.py
