+++ doc/apidoc/tests.rst
@@ -1,0 +1,12 @@
+Interesting tests
+=================
+
+These are here mostly for the sake of being cross-referenced if they
+provide useful examples.
+
+This page is therefore by no means comprehensive. If a given test
+class is listed here, only its interesting test cases are.
+
+.. autoclass:: anyblok_wms_base.core.goods.tests.test_containers.TestContainers
+
+   .. automethod:: test_override_recursion
